+++ compiler-core/src/type_/prelude.rs
@@ -244,6 +244,7 @@ pub fn build_prelude(ids: &UniqueIdGenerator) -> ModuleInterface {
         minimum_required_version: Version::new(0, 1, 0),
         type_aliases: HashMap::new(),
         documentation: Vec::new(),
+        contains_echo: false,
     };
 
     for t in PreludeType::iter() {