+++ compiler-core/src/type_/tests.rs
@@ -22,6 +22,7 @@ mod assert;
 mod assignments;
 mod conditional_compilation;
 mod custom_types;
+mod echo;
 mod errors;
 mod exhaustiveness;
 mod externals;
@@ -884,6 +885,7 @@ fn infer_module_type_retention_test() {
             minimum_required_version: Version::new(0, 1, 0),
             type_aliases: HashMap::new(),
             documentation: Vec::new(),
+            contains_echo: false,
         }
     );
 }